+++ packages/orq-rc/src/funcs/datasetsClear.ts
@@ -19,7 +19,6 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
-import * as errors from "../models/errors/index.js";
 import { OrqError } from "../models/errors/orqerror.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
@@ -31,7 +30,7 @@ import { Result } from "../types/fp.js";
  * Delete all datapoints
  *
  * @remarks
- * Delete all datapoints from a dataset. This action is irreversible.
+ * Deletes all datapoints from a dataset.
  */
 export function datasetsClear(
   client: OrqCore,
@@ -40,7 +39,6 @@ export function datasetsClear(
 ): APIPromise<
   Result<
     void,
-    | errors.HonoApiError
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -66,7 +64,6 @@ async function $do(
   [
     Result<
       void,
-      | errors.HonoApiError
       | OrqError
       | ResponseValidationError
       | ConnectionError
@@ -99,7 +96,7 @@ async function $do(
   const path = pathToFunc("/v2/datasets/{dataset_id}/clear")(pathParams);
 
   const headers = new Headers(compactMap({
-    Accept: "application/json",
+    Accept: "*/*",
   }));
 
   const secConfig = await extractSecurity(client._options.apiKey);
@@ -148,13 +145,8 @@ async function $do(
   }
   const response = doResult.value;
 
-  const responseFields = {
-    HttpMeta: { Response: response, Request: req },
-  };
-
   const [result] = await M.match<
     void,
-    | errors.HonoApiError
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -165,10 +157,9 @@ async function $do(
     | SDKValidationError
   >(
     M.nil(204, z.void()),
-    M.jsonErr(404, errors.HonoApiError$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
-  )(response, req, { extraFields: responseFields });
+  )(response, req);
   if (!result.ok) {
     return [result, { status: "complete", request: req, response }];
   }