+++ packages/orq-rc/src/funcs/datasetsRetrieveDatapoint.ts
@@ -11,6 +11,7 @@ import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
+import * as components from "../models/components/index.js";
 import {
   ConnectionError,
   InvalidRequestError,
@@ -18,7 +19,6 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
-import * as errors from "../models/errors/index.js";
 import { OrqError } from "../models/errors/orqerror.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
@@ -30,7 +30,7 @@ import { Result } from "../types/fp.js";
  * Retrieve a datapoint
  *
  * @remarks
- * Retrieves a datapoint object
+ * Retrieves a datapoint object.
  */
 export function datasetsRetrieveDatapoint(
   client: OrqCore,
@@ -38,8 +38,7 @@ export function datasetsRetrieveDatapoint(
   options?: RequestOptions,
 ): APIPromise<
   Result<
-    operations.RetrieveDatapointResponseBody,
-    | errors.HonoApiError
+    components.Datapoint1,
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -64,8 +63,7 @@ async function $do(
 ): Promise<
   [
     Result<
-      operations.RetrieveDatapointResponseBody,
-      | errors.HonoApiError
+      components.Datapoint1,
       | OrqError
       | ResponseValidationError
       | ConnectionError
@@ -153,13 +151,8 @@ async function $do(
   }
   const response = doResult.value;
 
-  const responseFields = {
-    HttpMeta: { Response: response, Request: req },
-  };
-
   const [result] = await M.match<
-    operations.RetrieveDatapointResponseBody,
-    | errors.HonoApiError
+    components.Datapoint1,
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -169,11 +162,10 @@ async function $do(
     | UnexpectedClientError
     | SDKValidationError
   >(
-    M.json(200, operations.RetrieveDatapointResponseBody$inboundSchema),
-    M.jsonErr(404, errors.HonoApiError$inboundSchema),
+    M.json(200, components.Datapoint1$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
-  )(response, req, { extraFields: responseFields });
+  )(response, req);
   if (!result.ok) {
     return [result, { status: "complete", request: req, response }];
   }