+++ packages/orq-rc/src/funcs/humanReviewSetsList.ts
@@ -35,7 +35,7 @@ export function humanReviewSetsList(
   options?: RequestOptions,
 ): APIPromise<
   Result<
-    Array<operations.GetV2HumanEvalSetsResponseBody>,
+    Array<operations.ResponseBody>,
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -60,7 +60,7 @@ async function $do(
 ): Promise<
   [
     Result<
-      Array<operations.GetV2HumanEvalSetsResponseBody>,
+      Array<operations.ResponseBody>,
       | OrqError
       | ResponseValidationError
       | ConnectionError
@@ -145,7 +145,7 @@ async function $do(
   const response = doResult.value;
 
   const [result] = await M.match<
-    Array<operations.GetV2HumanEvalSetsResponseBody>,
+    Array<operations.ResponseBody>,
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -155,10 +155,7 @@ async function $do(
     | UnexpectedClientError
     | SDKValidationError
   >(
-    M.json(
-      200,
-      z.array(operations.GetV2HumanEvalSetsResponseBody$inboundSchema),
-    ),
+    M.json(200, z.array(operations.ResponseBody$inboundSchema)),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req);