+++ packages/orq-rc/RELEASES.md
@@ -12539,3 +12539,13 @@ Based on:
 - [typescript v4.16.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.9 - packages/orq-rc
+
+## 2026-09-21 16:37:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.798.0 (2.938.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.10 - packages/orq-rc